feat: Update editor functionality and enhance homepage layout with new demo link

+++ src/app/dashboard/page.tsx
@@ -2,8 +2,7 @@
 
 import { useEffect, useState, useCallback, useMemo, Suspense } from 'react';
 import { useSession, signOut } from 'next-auth/react';
-import { useRouter } from 'next/navigation';
-import dynamic from 'next/dynamic';
+import { useRouter, useSearchParams } from 'next/navigation';
 import { ModeToggle } from '@/components/ui/themetoggle';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
@@ -21,6 +20,7 @@ import {
   Save,
   LogOut,
   Trash,
+  Link,
 } from 'lucide-react';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import Loader from '@/components/ui/loader';
@@ -44,12 +44,15 @@ const Dashboard = () => {
   const [isAlertDialogOpen, setIsAlertDialogOpen] = useState(false);
   const [nextNoteToSelect, setNextNoteToSelect] = useState<Note | null>(null);
   const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
+
   interface Note {
     id: string;
     title: string;
     content: string;
   }
 
+  const searchParams = useSearchParams();
+  const demo: boolean = searchParams.get('demo') === 'true';
   const [selectedNote, setSelectedNote] = useState<Note | null>(null);
   const [canEdit, setCanEdit] = useState<boolean>(true);
   const [searchQuery, setSearchQuery] = useState('');
@@ -57,10 +60,10 @@ const Dashboard = () => {
   const [selectedNoteContent, setSelectedNoteContent] = useState<string>('');
 
   useEffect(() => {
-    if (status === 'unauthenticated') {
+    if (status === 'unauthenticated' && !demo) {
       router.push('/login');
     }
-  }, [status, session, router]);
+  }, [status, session, router, demo]);
 
   useEffect(() => {
     const isMobileDevice = /Mobi|Android/i.test(navigator.userAgent);
@@ -70,7 +73,7 @@ const Dashboard = () => {
   }, []);
 
   useEffect(() => {
-    if (status === 'authenticated') {
+    if (status === 'authenticated' || demo) {
       fetchNotes();
     }
   }, [status]);
@@ -86,32 +89,58 @@ const Dashboard = () => {
     }
     const fuse = new Fuse(notesArray, {
       keys: ['title'],
-      threshold: 0.5,
+      threshold: 0.3,
     });
-    const result = fuse.search(searchQuery);
-    return result.map(({ item }) => item);
+    return fuse.search(searchQuery).map(({ item }) => item);
   }, [searchQuery, notesArray]);
 
   const toggleEdit = useCallback(() => {
-    setCanEdit((prevCanEdit) => !prevCanEdit);
+    setCanEdit((prev) => !prev);
   }, []);
 
   const fetchNotes = async () => {
+    if (demo) {
+      const demoNotes = [
+        {
+          id: '1',
+          title: 'Welcome to Memorious!',
+          content: `
+  <p class="welcome-paragraph">
+    You are currently in <strong>demo mode</strong>
+  </p>
+  <p>This mode is for demonstration purposes to showcase the app's features.</p>
+  <p>Here are some things you can do:</p>
+  <ul class>
+    <li><strong>Create</strong> new notes to organize your thoughts.</li>
+    <li><strong>Edit</strong> existing notes to refine your ideas.</li>
+    <li><strong>Delete</strong> notes when you no longer need them.</li>
+    <li>
+          Utilize <strong>Gemini AI</strong> to generate notes effortlessly from your prompts.</li>
+  </ul>
+  <p>
+      Have any feedback? Email us at
+      <a href="mailto:memorious.so@gmail.com" className="text-blue-500 underline">
+        memorious.so@gmail.com
+      </a>
+    </p>`,
+        },
+      ];
+      setNotesArray(demoNotes);
+      setSelectedNote(demoNotes[0]);
+      return;
+    }
     try {
       const response = await fetch(`/api/notes/user/${session?.user?.id}`, {
         method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-        },
+        headers: { 'Content-Type': 'application/json' },
       });
       if (!response.ok) {
         throw new Error('Failed to fetch notes');
       }
       const data = await response.json();
       setNotesArray(data && data.notes);
-      console.log('Fetched notes:', data.notes);
     } catch (error) {
-      console.error('Error fetching notes:', error);
+      console.error(error);
       toast.error('Failed to fetch notes.');
     }
   };
@@ -124,7 +153,7 @@ const Dashboard = () => {
     }
   }, [selectedNote]);
 
-  const handleSelectNote = (note: any) => {
+  const handleSelectNote = (note: Note) => {
     if (hasUnsavedChanges) {
       setNextNoteToSelect(note);
       setIsAlertDialogOpen(true);
@@ -157,16 +186,29 @@ const Dashboard = () => {
   const saveNote = useCallback(async () => {
     try {
       if (!selectedNote) return;
-
+      if (demo) {
+        if (selectedNote.id && !selectedNote.id.startsWith('temp-')) {
+          setNotesArray((prev) =>
+            prev.map((note) =>
+              note.id === selectedNote.id ? selectedNote : note
+            )
+          );
+        } else {
+          const newNote = { ...selectedNote, id: Date.now().toString() };
+          setSelectedNote(newNote);
+          setNotesArray((prev) => [
+            newNote,
+            ...prev.filter((x) => !x.id.startsWith('temp-')),
+          ]);
+        }
+        setHasUnsavedChanges(false);
+        return;
+      }
       if (selectedNote.id && !selectedNote.id.startsWith('temp-')) {
-        toast.loading('Saving...', {
-          id: 'save-note',
-        });
+        toast.loading('Saving...', { id: 'save-note' });
         const response = await fetch(`/api/notes/${selectedNote.id}`, {
           method: 'PUT',
-          headers: {
-            'Content-Type': 'application/json',
-          },
+          headers: { 'Content-Type': 'application/json' },
           body: JSON.stringify({
             title: selectedNote.title,
             content: selectedNote.content,
@@ -176,23 +218,16 @@ const Dashboard = () => {
           throw new Error('Failed to update note');
         }
         const data = await response.json();
-        setNotesArray(
-          notesArray.map((note) =>
-            note.id === data.note.id ? data.note : note
-          )
+        setNotesArray((prev) =>
+          prev.map((note) => (note.id === data.note.id ? data.note : note))
         );
         setHasUnsavedChanges(false);
         toast.dismiss('save-note');
       } else {
-        toast.loading('Saving...', {
-          id: 'create-note',
-        });
-
+        toast.loading('Saving...', { id: 'create-note' });
         const response = await fetch(`/api/notes/${session?.user?.id}`, {
           method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
+          headers: { 'Content-Type': 'application/json' },
           body: JSON.stringify({
             userId: session?.user?.id,
             title: selectedNote.title,
@@ -204,18 +239,18 @@ const Dashboard = () => {
         }
         const data = await response.json();
         setSelectedNote(data.note);
-        setNotesArray([
+        setNotesArray((prev) => [
           data.note,
-          ...notesArray.filter((note) => !note.id.startsWith('temp-')),
+          ...prev.filter((note) => !note.id.startsWith('temp-')),
         ]);
         setHasUnsavedChanges(false);
         toast.dismiss('create-note');
       }
     } catch (error) {
-      console.error('Error saving note:', error);
+      console.error(error);
       toast.error('Failed to save note.');
     }
-  }, [selectedNote, notesArray, hasUnsavedChanges]);
+  }, [selectedNote, notesArray, demo, session?.user?.id, hasUnsavedChanges]);
 
   const deleteNote = useCallback(() => {
     if (!selectedNote) return;
@@ -225,11 +260,15 @@ const Dashboard = () => {
   const handleDeleteNote = useCallback(async () => {
     if (!selectedNote) return;
     setIsDeleteDialogOpen(false);
-
-    toast.loading('Deleting...', {
-      id: 'delete-note',
-    });
-
+    if (demo) {
+      setNotesArray((prev) =>
+        prev.filter((note) => note.id !== selectedNote.id)
+      );
+      setSelectedNote(null);
+      setHasUnsavedChanges(false);
+      return;
+    }
+    toast.loading('Deleting...', { id: 'delete-note' });
     try {
       const response = await fetch(`/api/notes/${selectedNote.id}`, {
         method: 'DELETE',
@@ -237,17 +276,17 @@ const Dashboard = () => {
       if (!response.ok) {
         throw new Error('Failed to delete note');
       }
-      setNotesArray(notesArray.filter((note) => note.id !== selectedNote.id));
+      setNotesArray((prev) =>
+        prev.filter((note) => note.id !== selectedNote.id)
+      );
       setSelectedNote(null);
       setHasUnsavedChanges(false);
       toast.dismiss('delete-note');
     } catch (error) {
-      console.error('Error deleting note:', error);
-      toast.error('Failed to delete note.', {
-        id: 'delete-note',
-      });
+      console.error(error);
+      toast.error('Failed to delete note.', { id: 'delete-note' });
     }
-  }, [selectedNote, notesArray]);
+  }, [demo, selectedNote]);
 
   useEffect(() => {
     const handleKeyDown = (event: KeyboardEvent) => {
@@ -257,7 +296,6 @@ const Dashboard = () => {
           saveNote();
         }
       }
-      //delete key to delete note
       if (event.key.toLowerCase() === 'delete') {
         event.preventDefault();
         if (selectedNote) {
@@ -265,12 +303,8 @@ const Dashboard = () => {
         }
       }
     };
-
     window.addEventListener('keydown', handleKeyDown);
-
-    return () => {
-      window.removeEventListener('keydown', handleKeyDown);
-    };
+    return () => window.removeEventListener('keydown', handleKeyDown);
   }, [saveNote, deleteNote, hasUnsavedChanges, selectedNote]);
 
   useEffect(() => {
@@ -280,30 +314,25 @@ const Dashboard = () => {
         e.returnValue = '';
       }
     };
-
     window.addEventListener('beforeunload', handleBeforeUnload);
-
-    return () => {
-      window.removeEventListener('beforeunload', handleBeforeUnload);
-    };
+    return () => window.removeEventListener('beforeunload', handleBeforeUnload);
   }, [hasUnsavedChanges]);
 
-  const handleTitleChange = (e: { target: { value: any } }) => {
+  const handleTitleChange = (e: { target: { value: string } }) => {
     if (!selectedNote) return;
-    const updatedNote = { ...selectedNote, title: e.target.value };
-    setSelectedNote(updatedNote);
+    setSelectedNote({ ...selectedNote, title: e.target.value });
     setHasUnsavedChanges(true);
   };
 
   if (status === 'loading') {
     return (
-      <div className="h-screen w-screen flex items-center justify-center text text-3xl font-semibold">
+      <div className="h-screen w-screen flex items-center justify-center text-3xl font-semibold">
         <Loader />
       </div>
     );
   }
 
-  if (!session) {
+  if (!session && !demo) {
     return null;
   }
 
@@ -315,9 +344,7 @@ const Dashboard = () => {
             <Button
               variant={'secondary'}
               className="h-10 w-10"
-              onClick={() => {
-                setIsSidebarOpen(!isSidebarOpen);
-              }}
+              onClick={() => setIsSidebarOpen(!isSidebarOpen)}
             >
               {isSidebarOpen ? (
                 <PanelLeftClose className="h-5 w-5 flex-shrink-0" />
@@ -361,7 +388,7 @@ const Dashboard = () => {
         <div className="flex flex-1 overflow-hidden relative">
           <div
             className={`sidebar absolute lg:relative h-full ${
-              isSidebarOpen ? 'w-[80%] sm:w-[18rem] ' : 'w-0 border-r-0'
+              isSidebarOpen ? 'w-[80%] sm:w-[18rem]' : 'w-0 border-r-0'
             } overflow-hidden bg-background z-50 flex-shrink-0 border-r flex flex-col transition-all duration-500 text-nowrap flex-nowrap`}
           >
             <div className="flex flex-col gap-5 p-5 border-b border-border">
@@ -371,7 +398,7 @@ const Dashboard = () => {
                   className="justify-start w-full text-[1rem] px-3"
                   onClick={() => {
                     const newNote = {
-                      id: `temp-${Date.now()}`, // Assign a temporary unique ID
+                      id: `temp-${Date.now()}`,
                       title: '',
                       content: '',
                     };
@@ -416,7 +443,7 @@ const Dashboard = () => {
                           }`}
                         >
                           <FileText className="h-5 w-5 mr-2 stroke-2 flex-shrink-0" />
-                          <div className="text-nowrap  overflow-hidden">
+                          <div className="overflow-hidden">
                             {note.title || 'Untitled'}
                           </div>
                           <div
@@ -437,15 +464,18 @@ const Dashboard = () => {
                 <div className="bg-gradient-to-r from-transparent to-background w-7 h-full absolute right-0"></div>
                 <div className="bg-secondary p-3 rounded-md flex">
                   <div className="h-5 w-5 font-semibold text-[1.4rem] flex items-center justify-center">
-                    {session?.user?.name?.[0]?.toUpperCase() || ''}
+                    {/* {session?.user?.name?.[0]?.toUpperCase() || 'D'} */}
+                    {demo ? 'D' : session?.user?.name?.[0]?.toUpperCase()}
                   </div>
                 </div>
                 <div className="userinfo">
                   <div className="font-semibold text-[1rem] w-full">
-                    {session?.user?.name}
+                    {/* {session?.user?.name || (demo ? 'Demo User' : '')} */}
+                    {demo ? 'Demo User' : session?.user?.name}
                   </div>
                   <div className="text-muted-foreground text-[0.80rem] font-semibold">
-                    {session?.user?.email}
+                    {/* {session?.user?.email || (demo ? 'demo@example.com' : '')} */}
+                    {demo ? 'demo@example.com' : session?.user?.email}
                   </div>
                 </div>
               </div>
@@ -512,20 +542,17 @@ const Dashboard = () => {
             </div>
             {selectedNote ? (
               <div
-                className={`scrollarea  overflow-auto flex-1 flex transition-transform duration-500 py-16
-              ${
-                isSidebarOpen ? 'opacity-30 sm:opacity-100' : 'opacity-100'
-              } justify-center`}
+                className={`scrollarea  overflow-auto flex-1 flex transition-transform duration-500 py-16 ${
+                  isSidebarOpen ? 'opacity-30 sm:opacity-100' : 'opacity-100'
+                } justify-center`}
               >
-                <div
-                  className={`editorsection w-full md:w-[80%] lg:w-[760px] px-8`}
-                >
+                <div className="editorsection w-full md:w-[80%] lg:w-[760px] px-8">
                   <div
-                    className="title w-full h-fit font-bold text-3xl sm:text-5xl text-balance mb-2"
+                    className="title w-full h-fit font-bold text-3xl sm:text-5xl mb-2"
                     style={{ lineHeight: '1.2' }}
                   >
                     <TextAreaAutosize
-                      className="w-full titlefield bg-transparent outline-none text-balance font-bold text-3xl sm:text-5xl placeholder:text-muted-foreground resize-none"
+                      className="w-full bg-transparent outline-none font-bold text-3xl sm:text-5xl placeholder:text-muted-foreground resize-none"
                       value={selectedNote?.title || ''}
                       placeholder="Untitled"
                       readOnly={!canEdit}
@@ -549,7 +576,6 @@ const Dashboard = () => {
                 </div>
               </div>
             ) : (
-              // loader
               <div className="h-full w-full flex items-center justify-center">
                 <div>
                   <svg
@@ -558,9 +584,7 @@ const Dashboard = () => {
                     height="171"
                     viewBox="0 0 100 171"
                     fill="none"
-                    style={{
-                      height: '50%',
-                    }}
+                    style={{ height: '50%' }}
                   >
                     <path
                       d="M100 0H0V171L50 143L100 171V0Z"
@@ -579,11 +603,11 @@ const Dashboard = () => {
             <AlertDialogTitle>Unsaved Changes</AlertDialogTitle>
           </AlertDialogHeader>
           <AlertDialogDescription className="text-base font-medium">
-            You have unsaved changes. Do you want to save them before moving?
+            You have unsaved changes. Do you want to save them before moving on?
           </AlertDialogDescription>
           <AlertDialogFooter>
             <Button variant="outline" onClick={handleDiscardChanges}>
-              {`Don't Save`}
+              Don't Save
             </Button>
             <Button variant="secondary" onClick={handleSaveChanges}>
               Save
